feat: add per-group subpath exports (Strategy A modularization)
Create six group barrels (index.core, index.runtime, index.batch,
index.ws, index.abapgit, index.executors) so consumers can tree-shake
by importing '@mcp-abap-adt/adt-clients/core' etc. Root '.' re-exports
all six barrels — backward-compatible. Add exports map + typesVersions
to package.json. Smoke test verifies each barrel resolves its primary class.
