+++ src/index.abapgit.ts
@@ -1,0 +1,20 @@
+/**
+ * ADT Clients — abapGit barrel
+ * Covers: AdtAbapGitClient and all ./clients/abapGit/** types.
+ */
+
+export { AdtAbapGitClient } from './clients/AdtAbapGitClient';
+export type {
+  AbapGitStatus,
+  IAbapGitErrorLogEntry,
+  IAbapGitExternalRepoBranch,
+  IAbapGitExternalRepoCredentials,
+  IAbapGitExternalRepoInfo,
+  IAbapGitLinkArgs,
+  IAbapGitPullArgs,
+  IAbapGitPullResult,
+  IAbapGitRepoStatus,
+  IAbapGitUnlinkArgs,
+  IAdtAbapGitClient,
+  IAdtAbapGitClientOptions,
+} from './clients/abapGit/index';
