+++ src/index.executors.ts
@@ -1,0 +1,19 @@
+/**
+ * ADT Clients — executors barrel
+ * Covers: AdtExecutor and all ./executors/** types.
+ */
+
+export { AdtExecutor } from './clients/AdtExecutor';
+export type {
+  IClassExecuteWithProfilerOptions,
+  IClassExecuteWithProfilingOptions,
+  IClassExecuteWithProfilingResult,
+  IClassExecutionTarget,
+  IClassExecutor,
+  IExecutor,
+  IProgramExecuteWithProfilerOptions,
+  IProgramExecuteWithProfilingOptions,
+  IProgramExecuteWithProfilingResult,
+  IProgramExecutionTarget,
+  IProgramExecutor,
+} from './executors';
