+++ src/index.ws.ts
@@ -1,0 +1,14 @@
+/**
+ * ADT Clients — WebSocket barrel
+ * Covers: AdtClientsWS and DebuggerSessionClient.
+ */
+
+export { AdtClientsWS } from './clients/AdtClientsWS';
+export type {
+  DebuggerStepAction,
+  IDebuggerAttachParams,
+  IDebuggerGetVariablesParams,
+  IDebuggerListenParams,
+  IDebuggerStepParams,
+} from './clients/DebuggerSessionClient';
+export { DebuggerSessionClient } from './clients/DebuggerSessionClient';
